+++ src/content/etherscan/components/MainAddressLabel/index.tsx
@@ -1,13 +1,17 @@
 import { type FC } from 'react'
 import { Tooltip } from 'antd'
 
+import type { AddressLabel } from '@common/api/types'
+
 import { TokenSymbol } from '../../components'
 
 interface Props {
-  label: string
+  data: AddressLabel
 }
 
-const MainAddressLabel: FC<Props> = ({ label }) => {
+const MainAddressLabel: FC<Props> = ({
+  data: { label, implementLabel, implementAddress, implementLogo }
+}) => {
   return (
     <Tooltip
       title={
@@ -23,6 +27,15 @@ const MainAddressLabel: FC<Props> = ({ label }) => {
       <div className="badge bg-success bg-opacity-10 border border-success border-opacity-25 text-green-600 text-nowrap rounded-pill py-1.5 px-2">
         <TokenSymbol />
         {label}
+        {implementLabel?.trim() && (
+          <a
+            href={`/address/${implementAddress}`}
+            target="_blank"
+            rel="noreferrer"
+          >
+            {` ( ->`} <TokenSymbol logo={implementLogo} /> {implementLabel} )
+          </a>
+        )}
       </div>
     </Tooltip>
   )