+++ src/content/etherscan/page-scripts/tx.tsx
@@ -8,7 +8,7 @@ import {
   genAlternativeBlockExplorersBtn,
   genTxFortaAlertTip,
   genCopyIconBtn,
-  genTxPageAddressLabel,
+  genTxPageAddressLabels,
   genTransactionSummary
 } from '../feat-scripts'
 
@@ -22,13 +22,14 @@ const initTxPageScript = async (chain: string) => {
     enhancedLabels,
     txSummary
   } = await store.get('options')
+
   if (quick2Parsers && isSupportParsers(chain)) genQuick2ParsersBtn(chain)
   if (utc2locale) convertUTC2locale(SCAN_PAGES.TX.name)
   if (alternativeBlockExplorers)
     genAlternativeBlockExplorersBtn(SCAN_PAGES.TX.name, chain)
   if (txnFortaAlert) genTxFortaAlertTip(chain)
   if (showCopyIcon) genCopyIconBtn(SCAN_PAGES.TX.name)
-  if (enhancedLabels) genTxPageAddressLabel(chain)
+  if (enhancedLabels) genTxPageAddressLabels(chain)
   if (txSummary) genTransactionSummary(chain)
 }