+++ src/content/scans/feat-scripts/main-address-label.tsx
@@ -2,7 +2,7 @@ import type { CallbackResponse } from 'chrome-extension-core/lib/event'
 
 import { chromeEvent } from '@common/event'
 import type { AddressLabel } from '@common/api/types'
-import { GET_ADDRESS_LABELS } from '@common/constants'
+import { GET_IMPL_LABELS } from '@common/constants'
 import { createRoot } from 'react-dom/client'
 
 import { MainAddressLabel } from '../components'
@@ -16,7 +16,7 @@ const genMainAddressLabel = async (chain: string) => {
   if (!mainAddress) return
 
   await chromeEvent
-    .emit(GET_ADDRESS_LABELS, { chain: chain, addresses: [mainAddress] })
+    .emit(GET_IMPL_LABELS, { chain: chain, addresses: [mainAddress] })
     .then((res: CallbackResponse<AddressLabel[]> | undefined) => {
       if (res?.success && res.data.length) {
         const containerEl = document.querySelector(
@@ -28,7 +28,9 @@ const genMainAddressLabel = async (chain: string) => {
           labelRootEl.style.display = 'inline-block'
           labelRootEl.classList.add('mt-1')
           containerEl?.appendChild(labelRootEl)
-          createRoot(labelRootEl).render(<MainAddressLabel label={label} />)
+          createRoot(labelRootEl).render(
+            <MainAddressLabel data={res.data[0]} />
+          )
         }
       }
     })